+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 #### NEW FEATURES
 
+* add `BaselineCorrector()` widget  
+
 * `read_srs()` now allows extracting background interferogram or spectrum
 
 * add `download_nist_ir()` to download IR spectra from NIST/webbook